Pulsar searching files
splitSrch_v2. py is the file used to get just the upper half of the frequency range when analysing.
prep_fold.batch was a testing batch script when processing only one observation
uwl_search.batch is the batch script used to run pulsar searching on Garrawarla
run.sh  is the script uwl_search.batch runs, it conducts the pulsar searching.